+++ src/main/java/http/common/HttpMethod.java
@@ -1,4 +1,4 @@
-package http;
+package http.common;
 
 import java.util.HashMap;
 import java.util.Map;
