+++ src/main/java/http/controller/AbstractController.java
@@ -1,7 +1,7 @@
 package http.controller;
 
-import http.HttpRequest;
 import http.HttpResponse;
+import http.request.HttpRequest;
 
 public abstract class AbstractController implements Controller {
     @Override
