+++ src/main/java/http/controller/ResourcesController.java
@@ -1,7 +1,7 @@
 package http.controller;
 
-import http.HttpRequest;
 import http.HttpResponse;
+import http.request.HttpRequest;
 
 public class ResourcesController extends AbstractController {
     @Override
