+++ src/main/java/http/exception/InvalidRequestHeaderException.java
@@ -1,4 +1,4 @@
-package webserver;
+package http.exception;
 
 public class InvalidRequestHeaderException extends RuntimeException {
     public InvalidRequestHeaderException(String message) {
