+++ src/main/java/http/exception/InvalidResponseException.java
@@ -1,4 +1,4 @@
-package http;
+package http.exception;
 
 public class InvalidResponseException extends RuntimeException {
     public InvalidResponseException(String message) {
