+++ src/main/java/http/request/HttpRequestFactory.java
@@ -1,5 +1,6 @@
-package http;
+package http.request;
 
+import http.common.HeaderFields;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
