+++ src/test/java/http/common/HeaderFieldsTest.java
@@ -1,8 +1,6 @@
-package http;
+package http.common;
 
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class HeaderFieldsTest {
