+++ src/test/java/http/common/HttpMethodTest.java
@@ -1,9 +1,8 @@
-package http;
+package http.common;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class HttpMethodTest {
